+++ tests/dummy/app/routes/index.js
@@ -1,7 +1,8 @@
 import Ember from 'ember';
+import ENV from '../config/environment';
 
 export default Ember.Route.extend({
   redirect() {
-    window.location.replace('/tests?hidepassed');
+    window.location.replace(`${ENV.rootURL}docs`);
   }
 });