+++ tests/unit/validators/length-test.js
@@ -15,11 +15,8 @@ module('Unit | Validator | length');
 test('no options', function(assert) {
   assert.expect(1);
 
-  try {
-    result = validate(undefined, {});
-  } catch (e) {
-    assert.ok(true);
-  }
+  result = validate(undefined, {});
+  assert.ok(true);
 });
 
 test('allow blank', function(assert) {
@@ -91,6 +88,25 @@ test('max', function(assert) {
   assert.equal(processResult(result), 'This field is too long (maximum is 5 characters)');
 
   result = validate('test', cloneOptions(options));
+  assert.equal(processResult(result), true);
+});
+
+test('between', function(assert) {
+  assert.expect(3);
+
+  options = {
+    min: 1,
+    max: 5,
+    useBetweenMessage: true
+  };
+
+  result = validate('', cloneOptions(options));
+  assert.equal(processResult(result), 'This field must be between 1 and 5 characters');
+
+  result = validate('123456', cloneOptions(options));
+  assert.equal(processResult(result), 'This field must be between 1 and 5 characters');
+
+  result = validate('1234', cloneOptions(options));
   assert.equal(processResult(result), true);
 });