+++ CHANGELOG.md
@@ -1,4 +1,32 @@
-# Changelog
+Changelog
+=========
+
+## v1.0.2
+
+### Pull Requests
+
+- [#52](https://github.com/offirgolan/ember-validators/pull/52)  Use consistent semver of ember-require-module  *by [Lei Zhao](https://github.com/leizhao4)*
+
+## v1.0.1
+
+### Pull Requests
+
+- [#48](https://github.com/offirgolan/ember-validators/pull/48)  [BUGFIX] Remove assert in length validator  *by [Offir Golan](https://github.com/offirgolan)*
+
+## v1.0.0
+
+### Pull Requests
+
+- [#45](https://github.com/offirgolan/ember-validators/pull/45)  [FEATURE] Support Ember Changeset Validations  *by [Offir Golan](https://github.com/offirgolan)*
+
+## v0.2.0
+
+### Pull Requests
+
+- [#40](https://github.com/offirgolan/ember-validators/pull/40)  Remove unnecessary module caching  *by [Offir Golan](https://github.com/offirgolan)*
+- [#39](https://github.com/offirgolan/ember-validators/pull/39)  [FEATURE] allowNonTld & minTldLength in format validator  *by [Offir Golan](https://github.com/offirgolan)*
+- [#42](https://github.com/offirgolan/ember-validators/pull/42)  Expose parseDate function in the date validator  *by [Guillaume Gérard](https://github.com/GreatWizard)*
+- [#41](https://github.com/offirgolan/ember-validators/pull/41)  [BUGFIX] Use String.match instead of RegExp.test to support g flag  *by [Offir Golan](https://github.com/offirgolan)*
 
 ## v0.1.2